Python upgrade 3 12 (#30)
We initiated this project under Python 3.10 and now we are transitioning it to the latest version, Python 3.12, as of April 29, 2024. This upgrade is prompted by several compelling reasons:

1. **Syntactic Formalization of f-strings**: Python 3.12 introduces refined syntax for f-strings, enhancing readability and expressiveness in our codebase.
2. **Improved Error Messages**: With the upgrade, we gain access to more informative and precise error messages, which significantly aids debugging and maintenance.
3. **New Features Related to Type Hints**: Python 3.12 brings exciting enhancements to type hinting, allowing us to write more robust and self-documenting code.
4. **Long-Term Security Support**: By staying current with Python 3.12, we ensure that our project benefits from the latest security patches and updates, safeguarding our software against vulnerabilities.

This upgrade aligns our project with the latest advancements in the Python ecosystem, fostering code quality, maintainability, and security. We believe this transition will enhance the overall development experience and future-proof our project for continued success.