Integrate the SmartSocket for Live Price Data (#38)
### WebSocket Integration
We need a WebSocket data stream to collect the live data and store it in a database for backtesting. Also, we need this live data to test our strategies in the live market. Currently, we are integrating the SmartAPI WebSocket from angle one. 

**This PR contains**:
- Add the WebSocket client factory and protocol used to manage the lifecycle of the WebSocket connection
- Add the Base WebSocket class, `MarketDataWebSocket,` which handles the WebSocket interaction and events when connected to the WebSocket server. It is an abstract class, so it provides flexibility to the implementation class, like how they can subscribe and unsubscribe to the server with the tokens, etc.
- Add `SmartSocket` as the implementation class of the `MarketDataWebSocket` class. I will connect the `AngelOne` WebSocket server to get the data
- Add `WebSocketConnection` abstract base class to handle WebSocket connections
- Add `SmartSocketConnection ` as the implementation class of `WebSocketConnection` to handle the `SmartSocket` WebSocket connection
- Add `app/sockets/connect_to_websockets.py` script to connect to the multiple WebSockets with multiple instances using numerous threads
-  Added configs under `app/configs/websocket_connection` to manage the WebSocket configs

### Suspended the NSE tests
We have depended on the NSE to get the equity and option chain data, but recently they have changed their policies so we are not able to access their website through API. So all the tests that depended on the NSE website for the data were failing. We are suspending those tests ( comment on those tests) temporarily. We have to find some better alternatives for the data.